MVP Upload
fixed several bugs
Major bug with the date with the NASA picture of the day is now fixed. Had to subtract 24 hours from the current time so it doesn't fail anywhere in a different time zone.
now able to fetch data from SpaceX and NASA at the same time.
This commit does not include any design.